Move kernel.ts to kernels
